chore(infra): Dockerfile/Nx build 최적화
