+++ .vscode/settings.json
@@ -173,7 +173,7 @@
     "typescript": "javascript",
     "typescriptreact": "javascript"
   },
-
+  "cSpell.ignorePaths": ["pnpm-lock.yaml", "settings.json", "extensions.json", "snippets.code-snippets"],
   // 스펠링 검사에서 제외할 사용자 정의 단어 목록
   "cSpell.userWords": [
     "ShiwooMin",
@@ -359,7 +359,17 @@
     "apiuser",
     "loglevel",
     "ehthumbs",
-    "mjml"
+    "mjml",
+    "vercelignore",
+    "referer",
+    "letsencrypt",
+    "privkey",
+    "fullchain",
+    "ioredis",
+    "duckdns",
+    "certbot",
+    "snmaterial13",
+    "certonly"
   ],
 
   // Nx Console AI 규칙 사용
