+++ docs/guideline/04_의존성관리가이드.md
@@ -4,3 +4,15 @@ pnpm nx run @connectwon/api-contract:build --skip-nx-cache --verbose
 
 pnpm nx graph --file=e2e-graph.json
 pnpm nx connect
+pnpm outdated
+
+wsl --install -d Ubuntu
+
+certbot --version
+certbot 2.9.0
+
+sudo apt update
+sudo apt install ufw -y
+sudo ufw allow 80
+sudo ufw allow 443
+sudo ufw reload
