+++ docs/guideline/05_배포및운영가이드.md
@@ -8,3 +8,12 @@ docker builder prune -f
 
 # 3. 캐시 없이 완전 새로 빌드
 docker build --no-cache -f Dockerfile.api -t connectwon-api .
+
+start ms-settings:developers
+
+wsl -d Ubuntu
+
+sudo certbot certonly --standalone \
+  -d connectwon-web.duckdns.org \
+  -d connectwon-api.duckdns.org \
+  -d connectwon-admin.duckdns.org
